requirements.txt
annotated-types==0.7.0
anthropic==0.37.1
anyio==4.6.2.post1
asgiref==3.8.1
boto3==1.35.48
botocore==1.35.48
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==4.2.1
django-cors-headers==4.5.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
elevenlabs==1.12.1
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.1
idna==3.10
jiter==0.6.1
jmespath==1.0.1
openai==1.52.2
packaging==24.1
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
structlog==24.4.0
text-unidecode==1.3
tokenizers==0.20.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==1.26.20
websockets==13.1
celery==5.4.0
firebase-admin==6.6.0